refactor(ethereum_api): move config classes to models
