+++ packages/ethereum_api/lib/src/config/models/ethereum_address_config.dart
